created components and inputed and outputed values
