+++ src/app/app.component.html
@@ -1,16 +1,3 @@
-<p>Selected Monster: {{ monsterSelected.monsterName }}</p>
-<button type="button"
-  [disabled]="!monsterSelected.monsterName"
-  (click)="generatePower()">Generate Power</button>
-<p *ngIf="monsterSelected.monsterPower">{{ monsterSelected.monsterName }} has a power of {{ monsterSelected.monsterPower }}</p>
+<app-monster [heroPower]="heroPower"></app-monster>
 <hr>
-<p>All monsters:</p>
-<ul>
-  <li *ngFor="let monster of monsters" [ngStyle]="{'background-color': hero.power ? (monster.monsterPower > hero.power ? 'red': '') : ''}">
-    {{monster.monsterName}} has {{monster.monsterPower}} power!
-  </li>
-</ul>
-
-<input type="text" [(ngModel)]= "hero.name" placeholder="Enter Hero Name" *ngIf="!hero.power">
-<button type="button" (click)="generateHeroPower()" [disabled]="!hero.name" *ngIf="!hero.power">Select</button>
-<p *ngIf="hero.power">The power og hero is {{ hero.power }}</p>
+<app-hero (heroPower)="capturedHeroPower($event)"></app-hero>
