+++ src/app/app.component.ts
@@ -12,40 +12,14 @@ export class AppComponent implements OnInit {
  // monsterName:string = 'kaiju';
  // monsterPower: number;
 
-hero: any = {
-  name: '',
-  power: null
-};
+  heroPower: number;
 
- monsters: any[] = [
-   {
-
-     monsterName: 'kaiju',
-     monsterPower: null
-   },
-   {
-
-     monsterName: 'smaug',
-     monsterPower: null
-   }
- ];
-
- monsterSelected;
  ngOnInit() {
-   this.pickMonster();
+
 
  }
- pickMonster() {
-   const index = Math.floor(Math.random() * this.monsters.length);
-   this.monsterSelected = this.monsters[index];
- }
 
- generatePower() {
-  this.monsterSelected.monsterPower = Math.floor(Math.random() * 1000);
-
- }
- generateHeroPower() {
-  this.hero.power = Math.floor(Math.random() * 1000);
-  // console.log(this.hero.power);
+ capturedHeroPower(passedHeroPower: number) {
+  this.heroPower = passedHeroPower;
  }
 }
