+++ src/app/app.module.ts
@@ -3,10 +3,14 @@ import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 
 import { AppComponent } from './app.component';
+import { MonsterComponent } from './monster/monster.component';
+import { HeroComponent } from './hero/hero.component';
 
 @NgModule({
   declarations: [
-    AppComponent
+    AppComponent,
+    MonsterComponent,
+    HeroComponent
   ],
   imports: [
     BrowserModule,
