+++ src/app/hero/hero.component.html
@@ -1,0 +1,3 @@
+<input type="text" [(ngModel)]= "hero.name" placeholder="Enter Hero Name" *ngIf="!hero.power">
+<button type="button" (click)="generateHeroPower()" [disabled]="!hero.name" *ngIf="!hero.power">Select</button>
+<p *ngIf="hero.power">The power {{hero.name}} hero is {{ hero.power }}</p>
