+++ src/app/monster/monster.component.html
@@ -1,0 +1,13 @@
+<p>Selected Monster: {{ monsterSelected.monsterName }}</p>
+<button type="button"
+  [disabled]="!monsterSelected.monsterName"
+  (click)="generatePower()">Generate Power</button>
+<p *ngIf="monsterSelected.monsterPower">{{ monsterSelected.monsterName }} has a power of {{ monsterSelected.monsterPower }}</p>
+<p>All monsters:</p>
+<ul>
+  <li *ngFor="let monster of monsters" [ngStyle]="{'background-color': heroPower ? (monster.monsterPower > heroPower ? 'red': '') : ''}">
+    <!-- conditional style in ngFor that I may need later  -->
+    <!-- [ngStyle]="{'background-color': hero.power ? (monster.monsterPower > hero.power ? 'red': '') : ''}" -->
+    {{monster.monsterName}} has {{monster.monsterPower}} power!
+  </li>
+</ul>
